Factor out IntelliJ and compiler AST implementations into their own module
