+++ ast-compiler/README.md
@@ -1,0 +1,1 @@
+Intermediate representation of the Java AST based on Java's annotation processing APIs.
