+++ ast-compiler/build.gradle
@@ -1,0 +1,18 @@
+plugins {
+    id 'org.jetbrains.kotlin.jvm'
+    id 'org.jetbrains.dokka'
+}
+
+apply plugin: 'kotlin-kapt'
+
+sourceCompatibility = 1.8
+
+
+dependencies {
+    implementation deps.kotlin.stdlib
+    implementation deps.autoCommon
+
+    api project(':ast')
+}
+
+apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
