+++ ast-compiler/src/main/kotlin/motif/compiler/ir/CompilerClass.kt
