+++ compiler/build.gradle
@@ -6,6 +6,8 @@ plugins {
 sourceCompatibility = 1.8
 
 dependencies {
+    api project(':ast-compiler')
+
     implementation deps.kotlin.stdlib
     implementation deps.autoCommon
     implementation deps.commonsCodec
