+++ intellij/src/main/resources/gutter/childScope.png
