+++ intellij/src/main/resources/gutter/childScope@2x.png
