+++ intellij/src/main/resources/gutter/childScope@2x_dark.png
