+++ intellij/src/main/resources/gutter/childScope_dark.png
