+++ intellij/src/main/resources/gutter/parentScopes.png
