+++ intellij/src/main/resources/gutter/parentScopes@2x.png
