--- plugin/build.gradle
@@ -1,24 +1,0 @@
-plugins {
-    id "org.jetbrains.intellij" version "0.4.8"
-    id 'org.jetbrains.kotlin.jvm'
-}
-
-intellij {
-    version 'IC-2018.1'
-    plugins = []
-    pluginName 'MyPlugin'
-    updateSinceUntilBuild = false
-}
-
-group 'com.uber.motif'
-version '0.0.1' // Plugin version
-
-dependencies {
-    compile project(':lib')
-    compile project(':core')
-
-    compile deps.kotlin.reflection
-
-    testCompile deps.test.junit
-    testCompile deps.test.assertj
-}
