+++ settings.gradle
@@ -38,8 +38,10 @@ include 'samples:sample'
 include 'samples:sample-lib'
 include 'samples:sample-kotlin'
 include 'samples:dagger-comparison'
-include 'plugin'
+include 'intellij'
 include 'core'
 include 'ast'
+include 'ast-compiler'
+include 'ast-intellij'
 include 'it'
 include 'stub-compiler'
